refactor: Rename project and package name
Its purpose is to make the application more generic
