+++ app/src/main/java/com/globalaccelerex/mpossdksample/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.accelerex.swagger
+package com.globalaccelerex.mpossdksample
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupActionBarWithNavController
-import com.accelerex.swagger.databinding.ActivityMainBinding
+import com.globalaccelerex.mpossdksample.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
