+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Swagger</string>
+    <string name="app_name">Rex MPOS SDK Sample</string>
 </resources>
